backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0  # SQLite async support

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Utils
python-dateutil==2.8.2
pymupdf==1.24.0  # PDF text extraction

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
